requirements.txt
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.3
cron-descriptor==1.3.0
cryptography==40.0.2
cssselect2==0.7.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-import-export==3.2.0
django-timezone-field==5.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
fqdn==1.5.1
html5lib==1.1
idna==3.4
import-export==0.3.1
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jwt==1.3.1
kombu==5.2.4
lxml==4.9.2
mailtrap==2.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pypdf==3.8.1
pypng==0.20220715.0
pyrsistent==0.19.3
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.4
reportlab==3.6.13
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
tablib==3.4.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
uritools==4.0.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
widgetsnbextension==4.0.7
wkhtmltopdf==0.2
WTForms==3.0.1
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
